requirements.txt
langchain-astradb
langchain
langchain-huggingface
langchain-groq
datasets
pypdf
python-dotenv
flask

-e .